+++ lib/src/native/widget_state.dart
@@ -81,13 +81,13 @@ final class MapLibreMapStateNative extends State<MapLibreMap>
 
   @override
   Future<void> jumpTo({
-    required Position center,
+    Position? center,
     double? zoom,
     double? bearing,
     double? tilt,
   }) =>
       _hostApi.jumpTo(
-        center: center.toLngLat(),
+        center: center?.toLngLat(),
         zoom: zoom,
         bearing: bearing,
         pitch: tilt,
@@ -95,16 +95,18 @@ final class MapLibreMapStateNative extends State<MapLibreMap>
 
   @override
   Future<void> flyTo({
-    required Position center,
+    Position? center,
     double? zoom,
     double? bearing,
     double? tilt,
+    Duration duration = const Duration(seconds: 2),
   }) =>
       _hostApi.flyTo(
-        center: center.toLngLat(),
+        center: center?.toLngLat(),
         zoom: zoom,
         bearing: bearing,
         pitch: tilt,
+        durationMs: duration.inMilliseconds,
       );
 
   @override